refactor(layout): extract heading text predicates
Move heading-related text pattern checks into a dedicated layout helper. This separates caption, metadata, numbered-section, and body-fragment text heuristics from the geometry and font-size classification flow.
