+++ src/core/layout/headings.ts
@@ -1,5 +1,15 @@
 import type { LayoutBlock } from '../../types/index.js';
 import { median } from './geometry.js';
+import {
+  isDecimalSectionHeadingText,
+  isHeadingCandidateText,
+  isLetteredSectionHeadingText,
+  isLikelyBodyFragmentForLevel3,
+  isLikelyBodySentenceFragment,
+  isNumberedHeadingText,
+} from './headingText.js';
+
+export { isNumberedHeadingText } from './headingText.js';
 
 /** Min non-whitespace chars at the body font size required before low-tier
  *  (level 2 with structural support, or level 3) headings may fire. Pages
@@ -34,49 +44,6 @@ const COMPACT_LABEL_MAX_HEADING_CHARS = 4;
 const TALL_SIDE_LABEL_MIN_HEIGHT_PT = 80;
 const TALL_SIDE_LABEL_MAX_WIDTH_PT = 48;
 const TALL_SIDE_LABEL_MIN_ASPECT = 4;
-const CAPTION_IDENTIFIER_ATOM = '[A-Za-z\\p{N}０-９一二三四五六七八九十]+';
-const CAPTION_NUMBER_PATTERN = `${CAPTION_IDENTIFIER_ATOM}(?:(?:[.-]${CAPTION_IDENTIFIER_ATOM})|(?:-?\\(${CAPTION_IDENTIFIER_ATOM}\\)))*\\.?`;
-const CAPTION_HEADING_PATTERN = new RegExp(
-  `^\\s*(?:fig(?:ure)?\\.?|table|plate|図表|図|表)\\s*(${CAPTION_NUMBER_PATTERN})(?=\\s|[:：．、]|$)`,
-  'iu',
-);
-
-function isHeadingCandidateText(text: string): boolean {
-  const trimmed = text.trim();
-  if (!/[\p{L}\p{N}]/u.test(trimmed)) return false;
-  if (/^[\p{N}\s.-]{1,12}$/u.test(trimmed)) return false;
-  if (/^@[A-Za-z0-9_.-]{2,}$/u.test(trimmed)) return false;
-  if (isReferenceMetadataText(trimmed)) return false;
-  if (isCaptionHeadingText(trimmed)) return false;
-  return !/^[•●◦▪■‣]\s*/u.test(trimmed);
-}
-
-function isCaptionHeadingText(text: string): boolean {
-  return CAPTION_HEADING_PATTERN.test(text);
-}
-
-function isReferenceMetadataText(text: string): boolean {
-  return (
-    /\b(?:https?:\/\/|www\.|doi:|arxiv:)/iu.test(text) ||
-    /[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,}/iu.test(text) ||
-    /\bOMB\s+No\.?\s+\d/iu.test(text)
-  );
-}
-
-export function isNumberedHeadingText(text: string): boolean {
-  return /^\s*\d+(?:\.\d+)*\.?\s+\S/u.test(text);
-}
-
-function isDecimalSectionHeadingText(text: string): boolean {
-  return /^\s*\d+(?:\.\d+)+\.?\s+\p{L}/u.test(text.trim());
-}
-
-function isLetteredSectionHeadingText(text: string): boolean {
-  const trimmed = text.trim();
-  if (!/^[A-Z]\.\s+\p{Lu}/u.test(trimmed)) return false;
-  if (!/[.!?]$/u.test(trimmed)) return false;
-  return trimmed.split(/\s+/u).filter(Boolean).length >= 3;
-}
 
 function isTallNarrowSideLabel(block: LayoutBlock, lineCount: number): boolean {
   if (lineCount !== 1 || block.writingMode === 'vertical') return false;
@@ -90,26 +57,6 @@ function isCompactDiagramLabelText(text: string, nonWsChars: number, ratio: numb
   const trimmed = text.trim();
   if (isNumberedHeadingText(trimmed)) return false;
   return !/\s/u.test(trimmed);
-}
-
-function isLikelyBodyFragmentForLevel3(text: string): boolean {
-  if (isLikelyBodySentenceFragment(text)) return true;
-  const trimmed = text.trim();
-  if (isNumberedHeadingText(trimmed)) return false;
-  if (/[,;:]/u.test(trimmed)) return true;
-  return trimmed.split(/\s+/u).filter(Boolean).length > 7;
-}
-
-function isLikelyBodySentenceFragment(text: string): boolean {
-  const trimmed = text.trim();
-  if (isNumberedHeadingText(trimmed) || isLetteredSectionHeadingText(trimmed)) return false;
-  if (/^\p{Ll}/u.test(trimmed)) return true;
-  if (/\bet al\./iu.test(trimmed)) return true;
-  if (/[\p{L}\p{N}]-$/u.test(trimmed)) return true;
-  if (/[.!?。！？]$/u.test(trimmed)) return true;
-  const cjkChars = trimmed.match(/[\p{Script=Han}\p{Script=Hiragana}\p{Script=Katakana}]/gu)?.length ?? 0;
-  if (/[,;、。，；]/u.test(trimmed) && (trimmed.length > 24 || cjkChars >= 12)) return true;
-  return trimmed.split(/\s+/u).filter(Boolean).length > 16;
 }
 
 function isTopTitleCandidate(block: LayoutBlock, ratio: number, lineCount: number, nonWsChars: number): boolean {
